First refactoring of migration, still unfunctional
