+++ source/class/qx/tool/cli/commands/Lint.js
@@ -109,7 +109,7 @@ qx.Class.define("qx.tool.cli.commands.Lint", {
         }
         if (report.errorCount > 0 || report.warningCount > 0) {
           let outputFormat = this.argv.format || "codeframe";
-          
+
           // If there are too many errors, the pretty formatter is appallingly slow
           if (report.errorCount + report.warningCount > 150) {
             outputFormat = "compact";
@@ -174,7 +174,7 @@ qx.Class.define("qx.tool.cli.commands.Lint", {
           to: "@param $2 $1"
         });
       }
-      await this.migrate(null, replaceInFiles);
+      await qx.tool.migration.Utils.replaceInFiles(replaceInFiles);
     }
   }
 });
