+++ source/class/qx/tool/cli/commands/package/Install.js
@@ -37,7 +37,7 @@ qx.Class.define("qx.tool.cli.commands.package.Install", {
       return {
         command: "install [uri[@release_tag]]",
         describe: `installs the latest compatible release of package (as per Manifest.json). Use "-r <release tag>" or @<release tag> to install a particular release.
-        examples: 
+        examples:
            * qx package install name: Install latest published version
            * qx package install name@v0.0.2: Install version 0.0.2,
            * qx package install name@master: Install current master branch from github`,
@@ -244,8 +244,7 @@ qx.Class.define("qx.tool.cli.commands.package.Install", {
      * @private
      */
     __installFromRelease: async function(uri, tag_name, writeToManifest) {
-      let qooxdoo_version = await this.getUserQxVersion();
-
+      let qooxdoo_version = await qx.tool.config.Utils.getUserQxVersion(this.argv["block-global-framework"]);
       let {repo_name, package_path} = this.__getUriInfo(uri);
       if (!tag_name) {
         let cache = this.getCache();
@@ -308,7 +307,7 @@ qx.Class.define("qx.tool.cli.commands.package.Install", {
      * @private
      */
     __installFromTree: async function(uri, hash, writeToManifest) {
-      let qooxdoo_version = await this.getUserQxVersion();
+      let qooxdoo_version = await qx.tool.config.Utils.getUserQxVersion(this.argv["block-global-framework"]);
       if (this.argv.verbose) {
         qx.tool.compiler.Console.info(`>>> Installing '${uri}' from tree hash '${hash}' for qooxdoo version ${qooxdoo_version}`);
       }
@@ -330,7 +329,7 @@ qx.Class.define("qx.tool.cli.commands.package.Install", {
      * @private
      */
     async __installFromPath(uri, dir, writeToManifest=false) {
-      let qooxdoo_version = await this.getUserQxVersion();
+      let qooxdoo_version = await qx.tool.config.Utils.getUserQxVersion(this.argv["block-global-framework"]);
       if (this.argv.verbose) {
         qx.tool.compiler.Console.info(`>>> Installing '${uri}' from '${dir}' for qooxdoo version ${qooxdoo_version}`);
       }
@@ -440,7 +439,7 @@ qx.Class.define("qx.tool.cli.commands.package.Install", {
             break;
           case "qooxdoo-sdk":
           case "@qooxdoo/framework": {
-            let qxVer = await this.getUserQxVersion();
+            let qxVer = await qx.tool.config.Utils.getUserQxVersion(this.argv["block-global-framework"]);
             if (!semver.satisfies(qxVer, lib_range, {loose: true}) && this.argv.ignore) {
               throw new qx.tool.utils.Utils.UserError(
                 `Library '${lib_uri}' needs @qooxdoo/framework version ${lib_range}, found ${qxVer}`
